wrote a post about a paper

+++ build.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
-# nodemon --exec "./build.sh" . -e css,md,html
+set -eu
+
 
 input=${1:-personal-website}
 output=${2:-samuelstevens.github.io}
@@ -9,17 +10,27 @@ input_dir=${root}/${input}
 output_dir=${root}/${output}
 
 
+convert_single_md() {
+  md=$1
+  html=$(echo "$md" | sed -e "s/$input/$output/g" | sed "s/md/html/g")
+  dirname=$(echo "$md" | sed -e "s/$input/$output/g" | sed 's/md/html/g' | xargs dirname)
+  mkdir -p "$dirname"
+
+  # check if the file has references and needs '--citeproc'
+  if grep --quiet "^references:" "$md"; then
+    citeproc="--filter pandoc-citeproc"
+  else
+    citeproc=""
+  fi
+
+  pandoc --from markdown+backtick_code_blocks+link_attributes $citeproc --to html5 -s --template=template.html --ascii "$md" --out "$html"
+}
+
 full_convert_md () {
   for md in $(find ${input_dir} -name '*.md'); do
     if [[ $md != *README* ]]; then
       # echo $md | sed -e "s/$input/$output/g" | sed 's/md/html/g'
-
-      dirname=$(echo "$md" | sed -e "s/$input/$output/g" | sed 's/md/html/g' | xargs dirname)
-      mkdir -p "$dirname"
-
-      html=$(echo "$md" | sed -e "s/$input/$output/g" | sed "s/md/html/g")
-
-      pandoc --from markdown+backtick_code_blocks+link_attributes --to html5 -s --template=template.html --ascii "$md" --out "$html"
+      convert_single_md "$md"
     fi
   done
 }
@@ -27,15 +38,14 @@ full_convert_md () {
 latest_convert_md () {
   # finds the latest modified .md file.
   md=$(find ~/Development/personal-website -name '*.md' -exec ls -1t "{}" + | head -n 1)
-  html=$(echo $md | sed -e "s/$input/$output/g" | sed "s/md/html/g")
-  pandoc --from markdown+backtick_code_blocks+link_attributes --to html5 -s --template=template.html --ascii "$md" --out "$html"
+  convert_single_md "$md"
 }
 
 convert_html () {
   for html in $(find ${input_dir} -name '*.html'); do
     if [[ $html != *template* ]]; then
       # echo $html | sed -e "s/$input/$output/g"
-      cp $html $(echo $html | sed -e "s/$input/$output/g")
+      cp "$html" $(echo $html | sed -e "s/$input/$output/g")
     fi
   done
 }
@@ -53,6 +63,10 @@ build () {
     cat $css_file | tr -d '\n' | sed -E 's/[[:space:]]+/ /g' > ${output_dir}/css/$(basename $css_file)
   done
 
+  
+  mkdir -p "$output_dir/writing"
+  mkdir -p "$output_dir/projects"
+
   # gets the markdown and converts them to html
   latest_convert_md
 
@@ -60,12 +74,12 @@ build () {
   convert_html
 
   # copies js
-  mkdir -p $output_dir/js
+  mkdir -p "$output_dir/js"
   copy_js
 
   # copies images
-  mkdir -p $output_dir/images
-  cp -r $input_dir/images/* $output_dir/images
+  mkdir -p "$output_dir/images"
+  cp -r "$input_dir"/images/* $output_dir/images
 
   # copies my resume
   cp ~/Documents/Work/Sam\ Stevens\ Resume\ $(date +%Y).pdf ${output_dir}/resume.pdf
@@ -78,7 +92,7 @@ build () {
 }
 
 clean () {
-  rm -r $output_dir/* # shitty
+  rm -r "${output_dir:?}"/* # shitty
 }
 
 build
